% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{authorize_live}
\alias{authorize_live}
\title{Authorize sherwood for live environment}
\usage{
authorize_live(cache = FALSE)
}
\arguments{
\item{cache}{boolean, should the token be cached to file \code{.httr.oauth}?
Defaults to FALSE}
}
\value{
A Token2.0 reference class (RC) object
}
\description{
Wrapper around \code{httr::oauth2.0_token()}. Authorize sherwood app
for live environment in Saxo Banks OpenAPI, ie. enables trading
for real money.
}
\details{
The returned token object must be pased to subsequent API calls
using the \code{config} parameter in e.g. \code{httr::GET()}.

For more information about the oauth flow with \code{httr} see \href{https://github.com/r-lib/httr/blob/main/demo/oauth2-github.r}{this example}.
}
