% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cancel_all_orders}
\alias{cancel_all_orders}
\title{Cancel all orders
Call \code{get_orders()} and \code{cancel_order()} to cancel \emph{all} orders for an account}
\usage{
cancel_all_orders(token, account_key, ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{account_key}{character}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble, containing list of cancelled order ids
}
\description{
Cancel all orders
Call \code{get_orders()} and \code{cancel_order()} to cancel \emph{all} orders for an account
}
