% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_account_info}
\alias{get_account_info}
\title{Get accounts for user
Returns all accounts under a particular client to which the logged in user belongs.}
\usage{
get_account_info(token, ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble
}
\description{
Get accounts for user
Returns all accounts under a particular client to which the logged in user belongs.
}
\details{
See the \href{https://www.developer.saxo/openapi/referencedocs/port/v1/accounts/getaccounts/af56e3512758f8125dc6e5493d93c019}{Open API Documentation}
}
