% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_balance}
\alias{get_balance}
\title{Get cash balance for logged-in client}
\usage{
get_balance(token, ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
numeric, cash balance
}
\description{
Get cash balance for logged-in client
}
