% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_client_info}
\alias{get_client_info}
\title{Get client details}
\usage{
get_client_info(token, ...)
}
\arguments{
\item{token}{character}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble
}
\description{
Get client details
}
\details{
See the \href{https://www.developer.saxo/openapi/referencedocs/port/v1/clients/getclient/1499e70934cb99a0c9e70d53f9ad8f7d}{Open API Documentation}
}
