% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_cse_stocks}
\alias{get_cse_stocks}
\title{Get stock names and identifiers}
\usage{
get_cse_stocks(token)
}
\arguments{
\item{token}{character}
}
\value{
tibble
}
\description{
Get stock names and identifiers
}
