% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_info_prices}
\alias{get_info_prices}
\title{Get info (non-tradeable) prices for list of instruments}
\usage{
get_info_prices(token, uics, asset_type = "Stock", amount = 10000, ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{uics}{character, comma separated list of identifiers}

\item{asset_type}{character, defaults to "stock"}

\item{amount}{numeric, defaults to 10000}

\item{...}{parameters passed on to response_df()}
}
\value{
tibble with instruments (e.g. stocks)
}
\description{
Get info (non-tradeable) prices for list of instruments
}
