% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_instruments}
\alias{get_instruments}
\title{Get stock names and identifiers}
\usage{
get_instruments(token, exchange_id = "CSE", asset_type = "Stock", ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{exchange_id}{character, the Exchange to search, e.g. "CSE" for
Copenhagen Stock Exchange}

\item{asset_type}{character, type of asset, e.g. "Stock" or "StockOption"}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble
}
\description{
Get stock names and identifiers
}
