% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_optionspace}
\alias{get_optionspace}
\title{Get option space for a given root id}
\usage{
get_optionspace(token, live = FALSE, client_key, option_root_id)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{client_key}{character}

\item{option_root_id}{numeric, id of the required option root}
}
\value{
tibble
}
\description{
Returns the \verb{$SpecificOptions} element of the \verb{$OptionSpace}
element of the response object from a call to \code{contractoptionspaces}
endpoint.
}
