% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_orders}
\alias{get_orders}
\title{Get active orders for logged-in client}
\usage{
get_orders(token, ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble with active orders
}
\description{
Get active orders for logged-in client
}
