% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_schedule}
\alias{get_schedule}
\title{Get trading schedule for a given uic and asset type}
\usage{
get_schedule(token, uic, asset_type = "Stock", ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{uic}{character, instrument identifier}

\item{asset_type}{character, type of asset}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble with trading schedule
}
\description{
Get trading schedule for a given uic and asset type
}
