% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_stocks}
\alias{get_stocks}
\title{Get stock names and identifiers}
\usage{
get_stocks(token, exchange_id = "CSE")
}
\arguments{
\item{token}{character}
}
\value{
tibble
}
\description{
Get stock names and identifiers
}
