% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{make_subscription}
\alias{make_subscription}
\title{Create a price subscription on an instrument}
\usage{
make_subscription(token, live = FALSE, uic, asset_type)
}
\arguments{
\item{asset_type}{}
}
\value{

}
\description{
Create a price subscription on an instrument
}
