% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_id}
\alias{random_id}
\title{Generate a random id}
\usage{
random_id(no_chars = 10)
}
\arguments{
\item{no_chars, }{numeric, number of characters in randdom id}
}
\value{
character with random id
}
\description{
Useful for \code{context_id} and \code{reference_id} parameters
in \code{subscriptions} endpoint.
}
\examples{
random_id(5)

}
