% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel one or more orders}
\usage{
cancel_order(token, live = FALSE, account_key, order_ids)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{account_key}{character}

\item{order_ids}{character, one or more order ids}
}
\value{
tibble
}
\description{
Cancel one or more orders
}
\details{
See the \href{https://www.developer.saxo/openapi/referencedocs/trade/v2/orders/cancelorder/a1fd2fffa62f21901f23318f65fe8147}{Open API Documentation}
}
