% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_account_info}
\alias{get_account_info}
\title{Get accounts for user
Returns all accounts under a particular client to which the logged in user belongs.}
\usage{
get_account_info(token, live = FALSE, ...)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble
}
\description{
Get accounts for user
Returns all accounts under a particular client to which the logged in user belongs.
}
\details{
See the \href{https://www.developer.saxo/openapi/referencedocs/port/v1/accounts/getaccounts/af56e3512758f8125dc6e5493d93c019}{Open API Documentation}
}
