% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_balance}
\alias{get_balance}
\title{Get cash balance for logged-in client}
\usage{
get_balance(token, live = FALSE)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}
}
\value{
numeric, cash balance
}
\description{
Get cash balance for logged-in client
}
