% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_details}
\alias{get_details}
\title{Get detailed information for list of instruments, e.g. stocks}
\usage{
get_details(token, uics, asset_type = "Stock", ...)
}
\arguments{
\item{token}{character, authorization token, for demo environments
possibly a 24 hour token}

\item{uics}{character, comma separated list of identifiers}

\item{asset_type}{character, defaults to "stock"}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble with instruments (e.g. stocks)
}
\description{
Get detailed information for list of instruments, e.g. stocks
}
