% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_info_prices}
\alias{get_info_prices}
\title{Get info (non-tradeable) prices for list of instruments}
\usage{
get_info_prices(
  token,
  live = FALSE,
  uics,
  asset_type = "Stock",
  amount = 10000,
  ...
)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{uics}{character, comma separated list of identifiers}

\item{asset_type}{character, type of asset, e.g. "Stock" or "StockOption"}

\item{amount}{numeric, defaults to 10000}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble with instruments (e.g. stocks)
}
\description{
Get info (non-tradeable) prices for list of instruments
}
