% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_instruments}
\alias{get_instruments}
\title{Get stock names and identifiers}
\usage{
get_instruments(token, exchange_id = "CSE", asset_type = "Stock", ...)
}
\arguments{
\item{token}{character}

\item{exchange_id}{character}

\item{asset_type}{character}

\item{...}{parameters passed on to response_df()}
}
\value{
tibble
}
\description{
Get stock names and identifiers
}
