% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_schedule}
\alias{get_schedule}
\title{Get trading schedule for a given uic and asset type}
\usage{
get_schedule(token, live = FALSE, uic, asset_type = "Stock", ...)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{uic}{character, instrument identifier}

\item{asset_type}{character, type of asset}

\item{...}{parameters passed on to \code{response_df()}}
}
\value{
tibble with trading schedule
}
\description{
Get trading schedule for a given uic and asset type
}
