% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{make_subscription}
\alias{make_subscription}
\title{Create a price subscription on an instrument}
\usage{
make_subscription(token, live = FALSE, uic, asset_type)
}
\arguments{
\item{token}{either a character or a token2.0 reference class (RC) object
as returned by \code{httr::oauth2.0_token()}. Sim environment uses character,
(a '24 hour token'); live environment a token object.}

\item{live}{boolean, TRUE for live environment, i.e. real money.
Defaults to FALSE, i.e. simulation environment.}

\item{uic}{character, instrument identifier}

\item{asset_type}{character, type of asset, e.g. "Stock" or "StockOption"}
}
\value{
tibble
}
\description{
Create a price subscription on an instrument
}
