% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{response_df}
\alias{response_df}
\title{Convert HTTP response to dataframe}
\usage{
response_df(response, remove_constant_cols = FALSE)
}
\arguments{
\item{response}{A response object from {httr}}

\item{remove_constant_cols}{Logical. Should columns with constant values be removed? Defaults to TRUE.}
}
\value{
Tibble with response content
}
\description{
Convert HTTP response to dataframe
}
